+++ mobile/android/android-components/components/browser/toolbar/src/main/java/mozilla/components/browser/toolbar/behavior/BrowserToolbarBehavior.kt
@@ -6,19 +6,15 @@ package mozilla.components.browser.toolbar.behavior
 
 import android.content.Context
 import android.util.AttributeSet
-import android.view.Gravity
 import android.view.MotionEvent
 import android.view.View
 import androidx.annotation.VisibleForTesting
 import androidx.coordinatorlayout.widget.CoordinatorLayout
 import androidx.core.view.ViewCompat
-import com.google.android.material.snackbar.Snackbar
 import mozilla.components.browser.toolbar.BrowserToolbar
 import mozilla.components.concept.base.crash.CrashReporting
 import mozilla.components.concept.engine.EngineView
 import mozilla.components.support.ktx.android.view.findViewInHierarchy
-
-private const val SMALL_ELEVATION_CHANGE = 0.01f
 
 /**
  * Where the toolbar is placed on the screen.
@@ -36,7 +32,6 @@ enum class ToolbarPosition {
  *
  * This implementation will:
  * - Show/Hide the [BrowserToolbar] automatically when scrolling vertically.
- * - On showing a [Snackbar] position it above the [BrowserToolbar].
  * - Snap the [BrowserToolbar] to be hidden or visible when the user stops scrolling.
  */
 class BrowserToolbarBehavior(
@@ -130,14 +125,6 @@ class BrowserToolbarBehavior(
         return false // allow events to be passed to below listeners
     }
 
-    override fun layoutDependsOn(parent: CoordinatorLayout, child: BrowserToolbar, dependency: View): Boolean {
-        if (toolbarPosition == ToolbarPosition.BOTTOM && dependency is Snackbar.SnackbarLayout) {
-            positionSnackbar(child, dependency)
-        }
-
-        return super.layoutDependsOn(parent, child, dependency)
-    }
-
     override fun onLayoutChild(
         parent: CoordinatorLayout,
         child: BrowserToolbar,
@@ -179,23 +166,6 @@ class BrowserToolbarBehavior(
      */
     fun disableScrolling() {
         isScrollEnabled = false
-    }
-
-    @VisibleForTesting(otherwise = VisibleForTesting.PRIVATE)
-    internal fun positionSnackbar(child: View, snackbarLayout: Snackbar.SnackbarLayout) {
-        val params = snackbarLayout.layoutParams as CoordinatorLayout.LayoutParams
-
-        // Position the snackbar above the toolbar so that it doesn't overlay the toolbar.
-        params.anchorId = child.id
-        params.anchorGravity = Gravity.TOP or Gravity.CENTER_HORIZONTAL
-        params.gravity = Gravity.TOP or Gravity.CENTER_HORIZONTAL
-
-        snackbarLayout.layoutParams = params
-
-        // In order to avoid the snackbar casting a shadow on the toolbar we adjust the elevation of the snackbar here.
-        // We still place it slightly behind the toolbar so that it will not animate over the toolbar but instead pop
-        // out from under the toolbar.
-        snackbarLayout.elevation = child.elevation - SMALL_ELEVATION_CHANGE
     }
 
     @VisibleForTesting(otherwise = VisibleForTesting.PRIVATE)